--- src/tests/tests2-test.py
@@ -1,7 +1,0 @@
-import pytest
-
-
-class TestExample:
-    def test_addition(self):
-        assert 2 - 1 == 1
-
